# docs/swagger.yaml
basePath: /
definitions:
  apistruct.ApiResponse:
    properties:
      data: {}
      errCode:
        example: 0
        type: integer
      errDlt:
        example: ""
        type: string
      errMsg:
        example: success
        type: string
    type: object
...
  apistruct.SimpleSendMsgReq:
    properties:
      content:
        additionalProperties: true
        type: object
      contentType:
        example: 101
        type: integer
      ex:
        example: '{}'
        type: string
      groupID:
        example: group123
        type: string
      isOnlineOnly:
        example: false
        type: boolean
      notOfflinePush:
        example: false
        type: boolean
      recvID:
        example: user123
        type: string
      sendID:
        example: admin
        type: string
      sendTime:
        example: 1640995200000
        type: integer
      senderFaceURL:
        example: https://example.com/avatar.jpg
        type: string
      senderNickname:
        example: 管理员
        type: string
      senderPlatformID:
        example: 1
        type: integer
      sessionType:
        example: 1
        type: integer
    required:
    - content
    - contentType
    - sendID
    - sessionType
    type: object
  apistruct.SimpleSendMsgResp:
    properties:
      clientMsgID:
        example: client_msg_123
        type: string
      sendTime:
        example: 1640995200000
        type: integer
      serverMsgID:
        example: msg_123456
        type: string
    type: object
...
host: localhost:10002
info:
  contact:
    email: contact@openim.io
    name: OpenIM Support
    url: https://github.com/openimsdk/open-im-server
  description: |-
    OpenIM Server 提供即时通讯服务的 REST API 接口文档

    ## 认证方式
    大部分接口需要在 Header 中携带 token 进行认证：
    - Header名称: token
    - 获取方式: 调用 /auth/get_user_token 接口

    ## 接口规范
    - 所有接口均使用 POST 方法
    - Content-Type: application/json
    - 响应格式: JSON
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: OpenIM Server API
  version: "3.8"
paths:
...
  /msg/send_msg:
    post:
      consumes:
      - application/json
      description: 管理员发送消息到指定用户或群组
      parameters:
      - description: 消息内容
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/apistruct.SimpleSendMsgReq'
      produces:
      - application/json
      responses:
        "200":
          description: 成功响应
          schema:
            allOf:
            - $ref: '#/definitions/apistruct.ApiResponse'
            - properties:
                data:
                  $ref: '#/definitions/apistruct.SimpleSendMsgResp'
              type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/apistruct.ApiResponse'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/apistruct.ApiResponse'
        "403":
          description: 权限不足，仅管理员可发送
          schema:
            $ref: '#/definitions/apistruct.ApiResponse'
      security:
      - BearerToken: []
      summary: 发送消息
      tags:
      - 消息管理
...
schemes:
- http
- https
securityDefinitions:
  BearerToken:
    in: header
    name: token
    type: apiKey
swagger: "2.0"